+++ packages/api-derive/src/chain/subscribeFinalizedHeads.ts
@@ -41,7 +41,7 @@ export function subscribeFinalizedHeads (instanceId: string, api: DeriveApi): ()
         const endHash = prevHash;
         const startHash = header.parentHash;
 
-        prevHash = header.createdAtHash = header.hash;
+        prevHash = header.$createdAtHash = header.hash;
 
         return endHash === null || startHash.eq(endHash)
           ? of(header)