+++ packages/types-codec/src/abstract/Base.ts
@@ -9,17 +9,19 @@ import type { AnyJson, BareOpts, Codec, Inspect, IU8a, Registry } from '../types
  * @description A type extends the Base class, when it holds a value
  */
 export abstract class AbstractBase<T extends Codec> implements Codec {
-  public createdAtHash?: IU8a;
+  public $createdAtHash?: IU8a;
 
-  public readonly initialU8aLength?: number;
+  public $initialU8aLength?: number;
+
+  public $isStorageFallback?: boolean;
 
   public readonly registry: Registry;
 
   readonly #raw: T;
 
   protected constructor (registry: Registry, value: T, initialU8aLength?: number) {
+    this.$initialU8aLength = initialU8aLength;
     this.#raw = value;
-    this.initialU8aLength = initialU8aLength;
     this.registry = registry;
   }