+++ packages/types-codec/src/base/Enum.ts
@@ -185,13 +185,15 @@ function decodeEnum (registry: Registry, def: TypesDef, value?: unknown, index?:
 export class Enum implements IEnum {
   public readonly registry: Registry;
 
-  public createdAtHash?: IU8a;
+  public $createdAtHash?: IU8a;
+
+  public $initialU8aLength?: number;
+
+  public $isStorageFallback?: boolean;
 
   readonly #def: TypesDef;
 
   readonly #entryIndex: number;
-
-  readonly initialU8aLength?: number;
 
   readonly #indexes: number[];
 
@@ -217,8 +219,8 @@ export class Enum implements IEnum {
     this.#entryIndex = this.#indexes.indexOf(decoded.index);
     this.#raw = decoded.value;
 
-    if (this.#raw.initialU8aLength) {
-      this.initialU8aLength = 1 + this.#raw.initialU8aLength;
+    if (this.#raw.$initialU8aLength !== undefined) {
+      this.$initialU8aLength = 1 + this.#raw.$initialU8aLength;
     }
   }