+++ packages/types-codec/src/base/VecFixed.ts
@@ -25,8 +25,6 @@ function noopSetDefinition <T extends Codec> (d: CodecClass<T>): CodecClass<T> {
  * This manages codec arrays of a fixed length
  */
 export class VecFixed<T extends Codec> extends AbstractArray<T> {
-  readonly initialU8aLength?: number;
-
   #Type: CodecClass<T>;
 
   constructor (registry: Registry, Type: CodecClass<T> | string, length: number, value: Uint8Array | HexString | unknown[] = [] as unknown[], { definition, setDefinition = noopSetDefinition }: Options<T> = {}) {
@@ -34,7 +32,7 @@ export class VecFixed<T extends Codec> extends AbstractArray<T> {
 
     this.#Type = definition || setDefinition(typeToConstructor<T>(registry, Type));
 
-    this.initialU8aLength = (
+    this.$initialU8aLength = (
       isU8a(value)
         ? decodeU8aVec(registry, this, value, 0, this.#Type)
         : decodeVec(registry, this, value, 0, this.#Type)